requirements-dev.txt
jsonrpclib-pelix
pyeapi
netmiko
pyyaml
pytest==5.4.3
pylama==7.7.1
black==19.10b0
